progress
progress

+++ application/controllers/klik_a.php
@@ -142,7 +142,7 @@ class klik_a extends CI_Controller {
 			$this->form_validation->set_message('required','%s ngga boleh dikosongin');
 			$this->form_validation->set_message('is_unique','%s udah ada tuh, coba ganti yang lain deh');
 			$this->form_validation->set_message('max_length[20]','%s kepanjangan, maksimal 20 karakter');
-			$this->form_validation->set_message('max_length[2000]','%s kepanjangan, maksimal 20000 karakter');
+			$this->form_validation->set_message('max_length[2000]','%s kepanjangan, maksimal 2000 karakter');
 
 			if($this->form_validation->run() != false){
 				$nilai_master = array(
@@ -151,7 +151,8 @@ class klik_a extends CI_Controller {
 					'pst_mst_hak' => trim(strtoupper($this->input->post('t_pst_mst_hak'))),
 					'pst_mst_sts' => trim(strtoupper($this->input->post('t_pst_mst_sts'))),
 					'pst_mst_nm' => trim(strtoupper($this->input->post('t_pst_mst_nm'))),
-					'pst_mst_pswd' => MD5($this->input->post('t_pst_mst_pswd'))
+					'pst_mst_pswd' => MD5($this->input->post('t_pst_mst_pswd')),
+					'pst_mst_lock' => '0'
 				);
 
 				$kondisi = array ('pstprm' => $this->input->post('pstprm'));
@@ -236,7 +237,18 @@ class klik_a extends CI_Controller {
 
 	function hapus_peserta_ok($pstprm){
 		$kondisi = array ('pstprm' => $pstprm);
-		$this->m_db->hapus_data($kondisi,$this->TabelPesertaMaster);
+		$data['daftar_peserta_master'] = $this->m_db->ambil_data($kondisi,$this->TabelPesertaMaster)->result();
+		
+		foreach ($data['daftar_peserta_master'] as $pm) {
+				$t_pst_mst_lock = $pm->pst_mst_lock;
+			}
+		
+		if($t_pst_mst_lock == '0' ){
+			$this->m_db->hapus_data($kondisi,$this->TabelPesertaMaster);
+		} else {
+			$validasi_a2 = array('validasi_a2' => 'Pemakai ini lagi masuk, nda boleh hapus');
+			$this->session->set_userdata($validasi_a2);
+		}
 		redirect($this->KePilihanA2);
 	}
 
@@ -269,6 +281,7 @@ class klik_a extends CI_Controller {
 		$data['daftar_kelompok_master'] = $this->m_db->ambil_data($kondisi2,$this->TabelKelompokMaster)->result();
 		$data['daftar_peserta_master'] = $this->m_db->ambil_data_peserta($kondisi1,$this->TabelPesertaMaster)->result();
 		$data['daftar_info_level_1_sts'] = $this->m_db->ambil_data($kondisi3,$this->TabelInfoLevel1)->result();
+		
 		$this->load->view($this->FormA2,$data);
 
 		$this->kosong_operator_validasi();
@@ -288,12 +301,94 @@ class klik_a extends CI_Controller {
 
 		$this->kosong_operator_validasi();
 	}
+	
+	function proses_kelompok_ok(){
+		switch ($this->input->post('btnProses')){
+			case "UNDUH":
+				$nama_file = '_templateKelompok.csv';
+				$lokasi_file = file_get_contents(base_url().'berkas/unduh/_templateKelompok.csv');
+				force_download($nama_file,$lokasi_file);
+				break;
+			case "PROSES":
+				$berkas_sementara = $_FILES['t_kel_csv']['tmp_name'];
+				$berkas_aseli = $_FILES['t_kel_csv']['name'];
+				$ekstensi_berkas_aseli  = explode('.', $_FILES['t_kel_csv']['name']);
+				$ukuran_berkas_aseli = $_FILES['t_kel_csv']['size'];
+
+				if (strtolower(end($ekstensi_berkas_aseli)) === 'csv' && $ukuran_berkas_aseli>0) {
+					$baris = 0;
+					$proses_berkas = fopen($berkas_sementara, "r");
+
+					while ($kolom = fgetcsv($proses_berkas)) {
+						$baris++;
+						if ($baris == 1) continue; //untuk skip header csv
+						$isi_csv = array(
+							'kel_mst_sts' => trim(strtoupper($kolom[0])),
+							'kel_mst_kode' => trim(strtoupper($kolom[1])),
+							'kel_mst_subkode' => trim(strtoupper($kolom[2])),
+							'kel_mst_ket' => trim(strtoupper($kolom[3])),
+							'kel_mst_subket' => trim(strtoupper($kolom[4])),
+						);
+							
+						$this->m_db->tambah_data($isi_csv,$this->TabelKelompokMaster);
+					}
+					fclose($proses_berkas);
+				} else {
+					$validasi_a1 = array('validasi_a1' => 'Berkas yang di upload bukan file CSV atau isinya kosong...');
+					$this->session->set_userdata($validasi_a1);
+				}
+				redirect($this->KePilihanA1);
+				break;
+		}
+	}
+	
+	function proses_peserta_ok(){
+		switch ($this->input->post('btnProses')){
+			case "UNDUH":
+				$nama_file = '_templatePeserta.csv';
+				$lokasi_file = file_get_contents(base_url().'berkas/unduh/_templatePeserta.csv');
+				force_download($nama_file,$lokasi_file);
+				break;
+			case "PROSES":
+				$berkas_sementara = $_FILES['t_pst_csv']['tmp_name'];
+				$berkas_aseli = $_FILES['t_pst_csv']['name'];
+				$ekstensi_berkas_aseli  = explode('.', $_FILES['t_pst_csv']['name']);
+				$ukuran_berkas_aseli = $_FILES['t_pst_csv']['size'];
+
+				if (strtolower(end($ekstensi_berkas_aseli)) === 'csv' && $ukuran_berkas_aseli>0) {
+					$baris = 0;
+					$proses_berkas = fopen($berkas_sementara, "r");
+
+					while ($kolom = fgetcsv($proses_berkas)) {
+						$baris++;
+						if ($baris == 1) continue; //untuk skip header csv
+						$isi_csv = array(
+							'pst_mst_sts' => trim(strtoupper($kolom[0])),
+							'pst_mst_kode' => trim(strtoupper($kolom[1])),
+							'pst_mst_kel' => trim(strtoupper($kolom[2])),
+							'pst_mst_hak' => trim(strtoupper($kolom[3])),
+							'pst_mst_nm' => trim(strtoupper($kolom[4])),
+							'pst_mst_pswd' => trim(strtoupper($kolom[5])),
+							'pst_mst_lock' => '0'
+						);
+							
+						$this->m_db->tambah_data($isi_csv,$this->TabelPesertaMaster);
+					}
+					fclose($proses_berkas);
+				} else {
+					$validasi_a2 = array('validasi_a1' => 'Berkas yang di upload bukan file CSV atau isinya kosong...');
+					$this->session->set_userdata($validasi_a2);
+				}
+				redirect($this->KePilihanA2);
+				break;
+		}
+	}
 
 	function proses_rekening_ok(){
 		switch ($this->input->post('btnProses')){
 			case "UNDUH":
-				$nama_file = '_templetRek.csv';
-				$lokasi_file = file_get_contents(base_url().'berkas/unduh/_templetRek.csv');
+				$nama_file = '_templateRekening.csv';
+				$lokasi_file = file_get_contents(base_url().'berkas/unduh/_templateRekening.csv');
 				force_download($nama_file,$lokasi_file);
 				break;
 			case "PROSES":
@@ -310,16 +405,13 @@ class klik_a extends CI_Controller {
 						$baris++;
 						if ($baris == 1) continue; //untuk skip header csv
 						$isi_csv = array(
-							'rek_mst_kel' => trim(strtoupper($kolom[0])),
-							'rek_mst_gol' => trim(strtoupper($kolom[1])),
-							'rek_mst_sub_gol' => trim(strtoupper($kolom[2])),
-							'rek_mst_kode' => trim(strtoupper($kolom[3])),
-							'rek_mst_sub_kode' => trim(strtoupper($kolom[4])),
-							'rek_mst_sts' => trim(strtoupper($kolom[5])),
-							'rek_mst_ket_gol' => trim(strtoupper($kolom[6])),
-							'rek_mst_ket_sub_gol' => trim(strtoupper($kolom[7])),
-							'rek_mst_ket_kode' => trim(strtoupper($kolom[8])),
-							'rek_mst_ket_sub_kode' => trim(strtoupper($kolom[9]))
+							'rek_mst_sts' => trim(strtoupper($kolom[0])),
+							'rek_mst_kel' => trim(strtoupper($kolom[1])),
+							'rek_mst_gol' => trim(strtoupper($kolom[2])),
+							'rek_mst_sub_gol' => trim(strtoupper($kolom[3])),
+							'rek_mst_kode' => trim(strtoupper($kolom[4])),
+							'rek_mst_sub_kode' => trim(strtoupper($kolom[5])),
+							'rek_mst_ket_sub_kode' => trim(strtoupper($kolom[6]))
 						);
 							
 						$this->m_db->tambah_data($isi_csv,$this->TabelRekeningMaster);
@@ -349,16 +441,15 @@ class klik_a extends CI_Controller {
 	}
 
 	function cari_auto_kelompok_2_ok(){
-		$seperti1 = array('kel_mst_subkode' => $_GET['term']);
-		$seperti2 = array('kel_mst_subkode' => $_GET['extra']);
+		$seperti = array('kel_mst_subkode' => $_GET['term']);
 		$kondisi = array('kel_mst_kode' => $_GET['extra']);
 		$kelompok = array('kel_mst_subkode');
-		$data['daftar_kelompok_master'] = $this->m_db->ambil_data_seperti_kondisi($kondisi,$seperti1,$seperti2,$kelompok,$this->TabelKelompokMaster)->result();
+		$data['daftar_kelompok_master'] = $this->m_db->ambil_data_seperti_kondisi($kondisi,$seperti,$kelompok,$this->TabelKelompokMaster)->result();
 		
 		foreach($data['daftar_kelompok_master'] as $km){
 			$t_kel_mst_subkode[] = array(
 				'label' => substr($km->kel_mst_subkode,strlen($km->kel_mst_kode)),
-				'value' => $km->kel_mst_ket
+				'value' => $km->kel_mst_subket
 			);
 		}
 		echo json_encode($t_kel_mst_subkode);
